+++ src/custom/api/gnosisProtocol/errors/MetadataError.ts
