+++ src/custom/api/gnosisProtocol/errors/OperatorError.ts
