+++ src/custom/api/gnosisProtocol/errors/QuoteError.ts
