+++ src/custom/api/gnosisProtocol/operatorMock.ts
