+++ src/custom/state/lists/reducer/reducerMod.ts
@@ -22,7 +22,7 @@ import {
 } from 'state/lists/actions'
 import { SupportedChainId as ChainId } from 'constants/chains'
 import { getChainIdValues } from 'utils/misc'
-import { UnsupportedToken } from '@src/custom/api/gnosis'
+import { UnsupportedToken } from '@src/custom/api/gnosisProtocol'
 
 // Mod: change state shape - adds network map
 export type ListsStateByNetwork = {
