+++ src/custom/state/orders/reducer.ts
@@ -1,5 +1,5 @@
 import { createReducer, PayloadAction } from '@reduxjs/toolkit'
-import { OrderID } from '@src/custom/api/gnosis'
+import { OrderID } from '@src/custom/api/gnosisProtocol'
 import { SupportedChainId as ChainId } from 'constants/chains'
 import {
   addPendingOrder,
